refactor: adapt software card and move timeAgo to dateFn utils file
