+++ frontend/components/software/SoftwareCard.tsx
@@ -1,39 +1,15 @@
 import Link from 'next/link'
 import StarIcon from '@mui/icons-material/Star'
+import {getTimeAgoSince} from '../../utils/dateFn'
 
 export default function SoftwareCard({href,brand_name,short_statement,is_featured,updated_at}:
-  {href:string,brand_name:string,short_statement:string,is_featured:boolean,updated_at:string}) {
+  {href:string,brand_name:string,short_statement:string,is_featured:boolean,updated_at:string|null}) {
 
   const colors = is_featured ? 'bg-primary text-white' : 'bg-gray-200 text-gray-800'
   const today = new Date()
 
   function getInitals(){
     return brand_name.slice(0,2).toUpperCase()
-  }
-
-  function getTimeAgo(){
-    try{
-      const updated = new Date(updated_at)
-      if (today > updated){
-        const msDiff = today.getTime() - updated.getTime()
-        const hours = 1000 * 60 * 60
-        const hoursDiff = Math.floor(msDiff / hours)
-        if (hoursDiff > 24){
-          const daysDiff = Math.floor(hoursDiff/24)
-          if (daysDiff > 1) return `${hoursDiff} days ago`
-          return '1 day ago'
-        }else if (hoursDiff===1){
-          return '1 hour ago'
-        }else{
-          return `${hoursDiff} hours ago`
-        }
-      }else{
-        return 'right now'
-      }
-    }catch(e){
-      // on fail return nothing
-      return null
-    }
   }
 
   function renderFeatured(){
@@ -61,7 +37,7 @@ export default function SoftwareCard({href,brand_name,short_statement,is_feature
         </p>
         <div className="flex justify-between p-4 text-sm">
           <span className="last-update">
-            {getTimeAgo()}
+            {getTimeAgoSince(today,updated_at)}
           </span>
           {renderFeatured()}
         </div>
